+++ src/fetch.ts
@@ -112,7 +112,8 @@ export function fetchEventSource(input: RequestInfo, {
 
                 await onopen(response);
                 
-                await getBytes(response.body!, getLines(getMessages(id => {
+                await getBytes(response.body!, getLines(getMessages(onmessage,
+                   id => {
                     if (id) {
                         // store the id and send it back on the next retry:
                         headers[LastEventId] = id;
@@ -122,7 +123,7 @@ export function fetchEventSource(input: RequestInfo, {
                     }
                 }, retry => {
                     retryInterval = retry;
-                }, onmessage)));
+                })));
 
                 onclose?.();
                 dispose();