+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 3.0.0
+- Adjust parse.getMessages signature
 - Add node.js support
 - Upgrade to TypeScript 4.9
 - Add exports in package.json